New release
Add NetworkFX Ambient toggle button that switches profiles to
`netfx.profile` and spawns `eruption-netfx ambient`.
Add beginnings of a preferences dialog.
Fix a bug with the brightness slider, it should not get stuck anymore.
Fix a bug with the notifications code, only show notifications
if we have a valid monitor.